mlai/requirements_pytorch.txt
backports-zstd==1.3.0
dtale==3.18.2
duckdb==1.4.3
flask-wtf==1.2.2
fqdn==1.5.1
geopandas==1.1.2
imagesize==1.4.1
importlib-metadata==8.0.0
ipympl==0.9.8
isoduration==20.11.0
itables==2.6.2
jaraco.collections==5.1.0
jsonpointer==3.0.0
jupyter==1.1.1
lancedb==0.26.1
milvus-lite==2.5.1
opencv-contrib-python==4.10.0.84
paddleocr==3.3.2
paddlepaddle-gpu==2.6.2
pandera==0.28.0
pip-chill==1.0.3
pyclipper==1.4.0
pyjanitor==0.32.8
pypdfium2==5.3.0
python-bidi==0.6.7
python-dotenv==1.2.1
rfc3987-syntax==1.1.0
ruamel-yaml==0.19.1
sqlalchemy==2.0.45
surrealdb==1.0.7
tinycss2==1.4.0
tinydb==4.8.2
tomli==2.0.1
torchaudio==2.9.1+cu130
torchvision==0.24.1+cu130
tsfresh==0.21.1
uri-template==1.3.0
webcolors==25.10.0
xgboost==3.1.2
